refactor: Move ModelSpec from runtime to separatedir
