+++ cpp/tensorrt_llm/pybind/testing/modelSpecBinding.cpp
@@ -11,7 +11,7 @@
  */
 
 #include "modelSpecBinding.h"
-#include "tensorrt_llm/runtime/utils/modelSpec.h"
+#include "tensorrt_llm/testing/modelSpec.h"
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
