+++ cpp/tests/executor/executorMockTest.cpp
@@ -15,8 +15,8 @@
 #include "tensorrt_llm/batch_manager/trtGptModel.h"
 #include "tensorrt_llm/executor/executor.h"
 #include "tensorrt_llm/executor/types.h"
-#include "tensorrt_llm/runtime/utils/modelSpec.h"
 #include "tensorrt_llm/runtime/utils/mpiUtils.h"
+#include "tensorrt_llm/testing/modelSpec.h"
 #include "tests/utils/common.h"
 
 #include <gmock/gmock.h>
